chore(DI): update to v3.0.0
This is a breaking change: all classes that are instantiated via the DI
should be annotated (with one of @Injectable, @Component, @Decorator or
@Formatter).

Before this PR, the annotation were not enforced and the code could
break while switching to the static version of the DI because of a
missing annotation.

The impact on user code should be very limited as non-annotated classes
would not have worked before anyway. The error message is now clearer.

There is one corner case that might happen: by default the DI only
supports one annotation (@Injectable) and angular specific annotations
(@Component, @Decorator and @Formatter) are only added at the time a
dynamic application is created. If the code instantiates a module before
an application is created the angular specific annotations should be
added manually:

    assert(() {
      Module.classAnnotations.addAll[Component, Decorator, Formatter]);
      return true;
    });

    // Module can be instantiated here before the app is created

Closes #1475